feat: add latest events on start page
